+++ packages/core/src/routes/organization-scope/index.openapi.json
