+++ packages/core/src/routes/organization-scope/index.ts
@@ -3,9 +3,8 @@ import { OrganizationScopes } from '@logto/schemas';
 import koaQuotaGuard from '#src/middleware/koa-quota-guard.js';
 import SchemaRouter from '#src/utils/SchemaRouter.js';
 
+import { errorHandler } from '../organization/utils.js';
 import { type ManagementApiRouter, type RouterInitArgs } from '../types.js';
-
-import { errorHandler } from './utils.js';
 
 export default function organizationScopeRoutes<T extends ManagementApiRouter>(
   ...[
