+++ packages/core/src/routes/organization/application/index.openapi.json
@@ -8,6 +8,7 @@
   "paths": {
     "/api/organizations/{id}/applications": {
       "get": {
+        "tags": ["Dev feature"],
         "summary": "Get organization applications",
         "description": "Get applications associated with the organization.",
         "responses": {
@@ -17,6 +18,7 @@
         }
       },
       "post": {
+        "tags": ["Dev feature"],
         "summary": "Add organization application",
         "description": "Add an application to the organization.",
         "requestBody": {
@@ -42,6 +44,7 @@
         }
       },
       "put": {
+        "tags": ["Dev feature"],
         "summary": "Replace organization applications",
         "description": "Replace all applications associated with the organization with the given data.",
         "requestBody": {
@@ -69,6 +72,7 @@
     },
     "/api/organizations/{id}/applications/{applicationId}": {
       "delete": {
+        "tags": ["Dev feature"],
         "summary": "Remove organization application",
         "description": "Remove an application from the organization.",
         "responses": {
@@ -80,6 +84,7 @@
     },
     "/api/organizations/{id}/applications/{applicationId}/roles": {
       "get": {
+        "tags": ["Dev feature"],
         "summary": "Get organization application roles",
         "description": "Get roles associated with the application in the organization.",
         "responses": {
@@ -89,6 +94,7 @@
         }
       },
       "post": {
+        "tags": ["Dev feature"],
         "summary": "Add organization application role",
         "description": "Add a role to the application in the organization.",
         "requestBody": {
@@ -114,6 +120,7 @@
         }
       },
       "put": {
+        "tags": ["Dev feature"],
         "summary": "Replace organization application roles",
         "description": "Replace all roles associated with the application in the organization with the given data.",
         "requestBody": {
@@ -141,6 +148,7 @@
     },
     "/api/organizations/{id}/applications/{applicationId}/roles/{organizationRoleId}": {
       "delete": {
+        "tags": ["Dev feature"],
         "summary": "Remove organization application role",
         "description": "Remove a role from the application in the organization.",
         "responses": {
