+++ packages/core/src/routes/organization/index.ts
@@ -1,28 +1,21 @@
-import {
-  type OrganizationWithFeatured,
-  Organizations,
-  featuredUserGuard,
-  userWithOrganizationRolesGuard,
-} from '@logto/schemas';
+import { type OrganizationWithFeatured, Organizations, featuredUserGuard } from '@logto/schemas';
 import { yes } from '@silverhand/essentials';
 import { z } from 'zod';
 
-import { EnvSet } from '#src/env-set/index.js';
 import koaGuard from '#src/middleware/koa-guard.js';
 import koaPagination from '#src/middleware/koa-pagination.js';
 import koaQuotaGuard from '#src/middleware/koa-quota-guard.js';
-import { userSearchKeys } from '#src/queries/user.js';
 import SchemaRouter from '#src/utils/SchemaRouter.js';
 import { parseSearchOptions } from '#src/utils/search.js';
 
+import organizationInvitationRoutes from '../organization-invitation/index.js';
+import organizationRoleRoutes from '../organization-role/index.js';
+import organizationScopeRoutes from '../organization-scope/index.js';
 import { type ManagementApiRouter, type RouterInitArgs } from '../types.js';
 
-import applicationRoleRelationRoutes from './index.application-role-relations.js';
-import emailDomainRoutes from './index.jit.email-domains.js';
-import userRoleRelationRoutes from './index.user-role-relations.js';
-import organizationInvitationRoutes from './invitations.js';
-import organizationRoleRoutes from './roles.js';
-import organizationScopeRoutes from './scopes.js';
+import applicationRoutes from './application/index.js';
+import jitRoutes from './jit/index.js';
+import userRoutes from './user/index.js';
 import { errorHandler } from './utils.js';
 
 export default function organizationRoutes<T extends ManagementApiRouter>(
@@ -83,81 +76,9 @@ export default function organizationRoutes<T extends ManagementApiRouter>(
     }
   );
 
-  // MARK: Organization - user relation routes
-  router.addRelationRoutes(organizations.relations.users, undefined, {
-    disabled: { get: true },
-    hookEvent: 'Organization.Membership.Updated',
-  });
-
-  router.get(
-    '/:id/users',
-    koaPagination(),
-    koaGuard({
-      query: z.object({ q: z.string().optional() }),
-      params: z.object({ id: z.string().min(1) }),
-      response: userWithOrganizationRolesGuard.array(),
-      status: [200, 404],
-    }),
-    async (ctx, next) => {
-      const search = parseSearchOptions(userSearchKeys, ctx.guard.query);
-
-      const [totalCount, entities] = await organizations.relations.users.getUsersByOrganizationId(
-        ctx.guard.params.id,
-        ctx.pagination,
-        search
-      );
-
-      ctx.pagination.totalCount = totalCount;
-      ctx.body = entities;
-
-      return next();
-    }
-  );
-
-  // MARK: Organization - user role relation routes
-  router.post(
-    '/:id/users/roles',
-    koaGuard({
-      params: z.object({ id: z.string().min(1) }),
-      body: z.object({
-        userIds: z.string().min(1).array().nonempty(),
-        organizationRoleIds: z.string().min(1).array().nonempty(),
-      }),
-      status: [201, 422],
-    }),
-    async (ctx, next) => {
-      const { id } = ctx.guard.params;
-      const { userIds, organizationRoleIds } = ctx.guard.body;
-
-      await organizations.relations.usersRoles.insert(
-        ...organizationRoleIds.flatMap((roleId) =>
-          userIds.map((userId) => ({ organizationId: id, organizationRoleId: roleId, userId }))
-        )
-      );
-
-      ctx.status = 201;
-      return next();
-    }
-  );
-
-  userRoleRelationRoutes(router, organizations);
-
-  if (EnvSet.values.isDevFeaturesEnabled) {
-    // MARK: Organization - application relation routes
-    router.addRelationRoutes(organizations.relations.apps, undefined, {
-      hookEvent: 'Organization.Membership.Updated',
-    });
-
-    // MARK: Organization - application role relation routes
-    applicationRoleRelationRoutes(router, organizations);
-  }
-
-  // MARK: Just-in-time provisioning
-  emailDomainRoutes(router, organizations);
-  router.addRelationRoutes(organizations.jit.roles, 'jit/roles', { isPaginationOptional: true });
-  router.addRelationRoutes(organizations.jit.ssoConnectors, 'jit/sso-connectors', {
-    isPaginationOptional: true,
-  });
+  userRoutes(router, organizations);
+  applicationRoutes(router, organizations);
+  jitRoutes(router, organizations);
 
   // MARK: Mount sub-routes
   organizationRoleRoutes(...args);
